+++ lib/rspec-api/matchers.rb
@@ -3,7 +3,7 @@ require 'rspec-api/dsl/include_content_type'
 require 'rspec-api/dsl/have_prev_page_link'
 require 'rspec-api/dsl/be_a_jsonp'
 require 'rspec-api/dsl/be_sorted'
-require 'rspec-api/dsl/be_valid_json'
+require 'rspec-api/dsl/be_a_collection'
 require 'rspec-api/dsl/be_filtered'
 require 'rspec-api/dsl/have_attributes'
 require 'rspec-api/dsl/run_if' # should be the last, for metaprogramming